Logic class
Should run the overall logic in the game, currently only controlling input however it will tell every object when to update. updates should have a limit i.e. The player should not move as many pixels as as the game updates but should instead move a set amount of units per millisecond